requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.5.2 # Pinned to a recent stable version
Markdown==3.6 # Pinned to a recent stable version
gunicorn==22.0.0 # Pinned to a recent stable version 